+++ backend/src/main/java/com/votogether/domain/post/dto/response/vote/VoteOptionStatisticsResponse.java
@@ -1,4 +1,4 @@
-package com.votogether.domain.post.dto.response;
+package com.votogether.domain.post.dto.response.vote;
 
 import com.votogether.domain.member.entity.Gender;
 import java.util.Arrays;
