+++ backend/src/test/java/com/votogether/domain/post/controller/PostControllerTest.java
@@ -17,11 +17,11 @@ import com.votogether.domain.member.entity.Member;
 import com.votogether.domain.member.service.MemberService;
 import com.votogether.domain.post.dto.request.PostCreateRequest;
 import com.votogether.domain.post.dto.request.PostOptionCreateRequest;
-import com.votogether.domain.post.dto.response.PostDetailResponse;
+import com.votogether.domain.post.dto.response.detail.PostDetailResponse;
 import com.votogether.domain.post.dto.response.PostResponse;
-import com.votogether.domain.post.dto.response.VoteCountForAgeGroupResponse;
-import com.votogether.domain.post.dto.response.VoteDetailResponse;
-import com.votogether.domain.post.dto.response.VoteOptionStatisticsResponse;
+import com.votogether.domain.post.dto.response.vote.VoteCountForAgeGroupResponse;
+import com.votogether.domain.post.dto.response.detail.VoteDetailResponse;
+import com.votogether.domain.post.dto.response.vote.VoteOptionStatisticsResponse;
 import com.votogether.domain.post.dto.response.WriterResponse;
 import com.votogether.domain.post.entity.Post;
 import com.votogether.domain.post.entity.PostBody;
