+++ backend/src/test/java/com/votogether/domain/post/service/PostServiceTest.java
@@ -22,12 +22,11 @@ import com.votogether.domain.member.repository.MemberRepository;
 import com.votogether.domain.post.dto.request.PostOptionCreateRequest;
 import com.votogether.domain.post.dto.request.PostCreateRequest;
 import com.votogether.domain.post.dto.response.CategoryResponse;
-import com.votogether.domain.post.dto.response.PostDetailResponse;
-import com.votogether.domain.post.dto.response.PostOptionDetailResponse;
-import com.votogether.domain.post.dto.response.PostOptionResponse;
+import com.votogether.domain.post.dto.response.detail.PostDetailResponse;
+import com.votogether.domain.post.dto.response.detail.PostOptionDetailResponse;
 import com.votogether.domain.post.dto.response.PostResponse;
-import com.votogether.domain.post.dto.response.VoteDetailResponse;
-import com.votogether.domain.post.dto.response.VoteOptionStatisticsResponse;
+import com.votogether.domain.post.dto.response.detail.VoteDetailResponse;
+import com.votogether.domain.post.dto.response.vote.VoteOptionStatisticsResponse;
 import com.votogether.domain.post.dto.response.WriterResponse;
 import com.votogether.domain.post.entity.Post;
 import com.votogether.domain.post.entity.PostBody;
@@ -46,7 +45,6 @@ import com.votogether.fixtures.CategoryFixtures;
 import com.votogether.fixtures.MemberFixtures;
 import jakarta.persistence.EntityManager;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
